requirements.txt
appnope==0.1.0
asana==0.6.5
backports.shutil-get-terminal-size==1.0.0
certifi==2017.11.5
chardet==3.0.4
cookies==2.2.1
decorator==4.1.2
enum34==1.1.6
funcsigs==1.0.2
idna==2.6
ipython==5.5.0
ipython-genutils==0.2.0
mock==2.0.0
numpy==1.13.3
oauthlib==2.0.6
pandas==0.21.1
pathlib2==2.3.0
pbr==3.1.1
pexpect==4.3.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py==1.5.2
Pygments==2.2.0
pytest==2.8.7
python-dateutil==2.6.1
pytz==2017.3
requests==2.14.2
requests-oauthlib==0.6.2
responses==0.5.1
scandir==1.6
simplegeneric==0.8.1
six==1.10.0
todoist-python==7.0.18
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
